+++ concordium-sdk/src/main/java/com/concordium/sdk/crypto/RawKey.java
@@ -1,0 +1,9 @@
+package com.concordium.sdk.crypto;
+
+/**
+ * Interface for keys exposed in a raw format i.e. a byte array.
+ */
+public interface RawKey {
+
+    byte[] getRawBytes();
+}