+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/IdJsonSerializer.java
@@ -1,0 +1,17 @@
+package com.concordium.sdk.responses;
+
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+
+import java.io.IOException;
+
+/**
+ * Serializing ids to json.
+ */
+public class IdJsonSerializer extends JsonSerializer<ID> {
+    @Override
+    public void serialize(ID value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
+        gen.writeNumber(value.getId());
+    }
+}