+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blockitemsummary/AccountTransactionDetails.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.responses.blockitemsummary;
 
+import com.concordium.grpc.v2.AccountTransactionEffects;
 import com.concordium.sdk.responses.modulelist.ModuleRef;
 import com.concordium.sdk.responses.smartcontracts.ContractTraceElement;
 import com.concordium.sdk.responses.transactionstatus.*;
@@ -13,8 +14,13 @@ import java.util.stream.Collectors;
 
 /**
  * Details of a transaction with a sender account.
+ *
+ * Users should always check whether the {@link AccountTransactionDetails} is {@link AccountTransactionDetails#isSuccessful()} or not.
+ * If this returns false, then one should consult the {@link AccountTransactionDetails#rejectReason} for why the transaction failed.
+ * If the transaction was successful then first check the type via {@link AccountTransactionDetails#getType()} and use the corresponding
+ * getter for getting the concrete event.
  */
-@EqualsAndHashCode
+@EqualsAndHashCode(doNotUseGetters = true)
 @ToString(doNotUseGetters = true)
 @Builder
 @Getter
@@ -306,7 +312,6 @@ public class AccountTransactionDetails {
                 .sender(sender)
                 .cost(CCDAmount.from(tx.getCost()))
                 .successful(true);
-
         val effects = tx.getEffects();
         switch (effects.getEffectCase()) {
             case NONE:
@@ -315,10 +320,14 @@ public class AccountTransactionDetails {
                         .rejectReason(RejectReasonType.from(effects.getNone().getRejectReason()));
                 break;
             case MODULE_DEPLOYED:
-                detailsBuilder.moduleDeployed(ModuleRef.from(effects.getModuleDeployed().getValue().toByteArray()));
+                detailsBuilder
+                        .type(TransactionResultEventType.MODULE_DEPLOYED)
+                        .moduleDeployed(ModuleRef.from(effects.getModuleDeployed().getValue().toByteArray()));
                 break;
             case CONTRACT_INITIALIZED:
-                detailsBuilder.contractInitialized(ContractInitializedResult.from(effects.getContractInitialized()));
+                detailsBuilder
+                        .type(TransactionResultEventType.CONTRACT_INITIALIZED)
+                        .contractInitialized(ContractInitializedResult.from(effects.getContractInitialized()));
                 break;
             case CONTRACT_UPDATE_ISSUED:
                 val updateEvents = effects
@@ -327,55 +336,85 @@ public class AccountTransactionDetails {
                         .stream()
                         .map(ContractTraceElement::from)
                         .collect(Collectors.toList());
-                detailsBuilder.contractUpdated(updateEvents);
+                detailsBuilder
+                        .type(TransactionResultEventType.CONTRACT_UPDATED)
+                        .contractUpdated(updateEvents);
                 break;
             case ACCOUNT_TRANSFER:
-                if (effects.getAccountTransfer().hasMemo()) {
-                    break;
-                }
-                detailsBuilder.accountTransfer(TransferredResult.from(effects.getAccountTransfer(), sender));
+                detailsBuilder
+                        .type(TransactionResultEventType.TRANSFERRED)
+                        .accountTransfer(TransferredResult.from(effects.getAccountTransfer(), sender));
+
                 break;
             case BAKER_ADDED:
-                detailsBuilder.bakerAdded(BakerAddedResult.from(effects.getBakerAdded()));
+                detailsBuilder
+                        .type(TransactionResultEventType.BAKER_ADDED)
+                        .bakerAdded(BakerAddedResult.from(effects.getBakerAdded()));
                 break;
             case BAKER_REMOVED:
-                detailsBuilder.bakerRemoved(BakerRemovedResult.from(effects.getBakerRemoved(), sender));
+                detailsBuilder
+                        .type(TransactionResultEventType.BAKER_REMOVED)
+                        .bakerRemoved(BakerRemovedResult.from(effects.getBakerRemoved(), sender));
                 break;
             case BAKER_STAKE_UPDATED:
-                detailsBuilder.bakerStakeUpdated(BakerStakeUpdated.from(effects.getBakerStakeUpdated(), sender));
+                detailsBuilder
+                        .type(TransactionResultEventType.BAKER_STAKE_UPDATED)
+                        .bakerStakeUpdated(BakerStakeUpdated.from(effects.getBakerStakeUpdated(), sender));
                 break;
             case BAKER_RESTAKE_EARNINGS_UPDATED:
-                detailsBuilder.bakerRestakeEarningsUpdated(BakerSetRestakeEarningsResult.from(effects.getBakerRestakeEarningsUpdated(), sender));
+                detailsBuilder
+                        .type(TransactionResultEventType.BAKER_SET_RESTAKE_EARNINGS)
+                        .bakerRestakeEarningsUpdated(BakerSetRestakeEarningsResult.from(effects.getBakerRestakeEarningsUpdated(), sender));
                 break;
             case BAKER_KEYS_UPDATED:
-                detailsBuilder.bakerKeysUpdated(BakerKeysUpdatedResult.from(effects.getBakerKeysUpdated(), sender));
+                detailsBuilder
+                        .type(TransactionResultEventType.BAKER_KEYS_UPDATED)
+                        .bakerKeysUpdated(BakerKeysUpdatedResult.from(effects.getBakerKeysUpdated(), sender));
                 break;
             case ENCRYPTED_AMOUNT_TRANSFERRED:
-                detailsBuilder.encryptedTransfer(EncryptedTransferResult.from(effects.getEncryptedAmountTransferred()));
+                detailsBuilder
+                        .type(TransactionResultEventType.ENCRYPTED_TRANSFER)
+                        .encryptedTransfer(EncryptedTransferResult.from(effects.getEncryptedAmountTransferred()));
                 break;
             case TRANSFERRED_TO_ENCRYPTED:
-                detailsBuilder.addedToEncryptedBalance(EncryptedSelfAmountAddedResult.from(effects.getTransferredToEncrypted()));
+                detailsBuilder
+                        .type(TransactionResultEventType.ENCRYPTED_SELF_AMOUNT_ADDED)
+                        .addedToEncryptedBalance(EncryptedSelfAmountAddedResult.from(effects.getTransferredToEncrypted()));
                 break;
             case TRANSFERRED_TO_PUBLIC:
-                detailsBuilder.removedFromEncryptedBalance(EncryptedAmountsRemovedResult.from(effects.getTransferredToPublic().getRemoved()));
+                detailsBuilder
+                        .type(TransactionResultEventType.ENCRYPTED_AMOUNTS_REMOVED)
+                        .removedFromEncryptedBalance(EncryptedAmountsRemovedResult.from(effects.getTransferredToPublic().getRemoved()));
                 break;
             case TRANSFERRED_WITH_SCHEDULE:
-                detailsBuilder.transferredWithSchedule(TransferredWithScheduleResult.from(effects.getTransferredWithSchedule(), sender));
+                detailsBuilder
+                        .type(TransactionResultEventType.TRANSFERRED_WITH_SCHEDULE)
+                        .transferredWithSchedule(TransferredWithScheduleResult.from(effects.getTransferredWithSchedule(), sender));
                 break;
             case CREDENTIAL_KEYS_UPDATED:
-                detailsBuilder.credentialKeysUpdated(CredentialKeysUpdatedResult.from(effects.getCredentialKeysUpdated()));
+                detailsBuilder
+                        .type(TransactionResultEventType.CREDENTIAL_KEYS_UPDATED)
+                        .credentialKeysUpdated(CredentialKeysUpdatedResult.from(effects.getCredentialKeysUpdated()));
                 break;
             case CREDENTIALS_UPDATED:
-                detailsBuilder.credentialsUpdated(CredentialsUpdatedResult.from(effects.getCredentialsUpdated(), sender));
+                detailsBuilder
+                        .type(TransactionResultEventType.CREDENTIALS_UPDATED)
+                        .credentialsUpdated(CredentialsUpdatedResult.from(effects.getCredentialsUpdated(), sender));
                 break;
             case DATA_REGISTERED:
-                detailsBuilder.dataRegistered(DataRegisteredResult.from(effects.getDataRegistered()));
+                detailsBuilder
+                        .type(TransactionResultEventType.DATA_REGISTERED)
+                        .dataRegistered(DataRegisteredResult.from(effects.getDataRegistered()));
                 break;
             case BAKER_CONFIGURED:
-                detailsBuilder.bakerConfigured(BakerConfigured.from(effects.getBakerConfigured(), sender));
+                detailsBuilder
+                        .type(TransactionResultEventType.BAKER_CONFIGURED)
+                        .bakerConfigured(BakerConfigured.from(effects.getBakerConfigured(), sender));
                 break;
             case DELEGATION_CONFIGURED:
-                detailsBuilder.delegatorConfigured(DelegatorConfigured.from(effects.getDelegationConfigured(), sender));
+                detailsBuilder
+                        .type(TransactionResultEventType.DELEGATION_CONFIGURED)
+                        .delegatorConfigured(DelegatorConfigured.from(effects.getDelegationConfigured(), sender));
                 break;
             case EFFECT_NOT_SET:
                 throw new IllegalArgumentException("Unrecognized effect.");